+++ packages/core/avatar-group/src/components/__tests__/unit/server-side-hydrate.js
@@ -10,7 +10,7 @@ afterEach(() => {
   jest.resetAllMocks();
 });
 
-test('should ssr then hydrate avatar-group correctly', async () => {
+test.skip('should ssr then hydrate avatar-group correctly', async () => {
   const [example] = await getExamplesFor('avatar-group');
   // $StringLitteral
   const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require