+++ packages/core/tree/src/components/Tree/__tests__/unit/server-side-hydrate.js
@@ -9,8 +9,8 @@ jest.spyOn(global.console, 'error');
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate tree correctly', async () => {
+// TODO: https://ecosystem.atlassian.net/browse/AK-5988
+test.skip('should ssr then hydrate tree correctly', async () => {
   const [example] = await getExamplesFor('tree');
   // $StringLitteral
   const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require