+++ packages/editor/editor-mobile-bridge/src/editor/mobile-editor-element.tsx
@@ -57,7 +57,7 @@ class EditorWithState extends Editor {
   }
 }
 
-export default function mobileEditor(props) {
+export default function mobileEditor(props: any) {
   return (
     <EditorWithState
       appearance="mobile"