+++ packages/navigation/atlassian-switcher/examples/45-generic-switcher-with-xflow.tsx
@@ -1,11 +1,11 @@
-import React, { Component } from 'react';
+import * as React from 'react';
 import Button from '@atlaskit/button';
 import Drawer from '@atlaskit/drawer';
-import { mockEndpoints } from './helpers/mock-endpoints';
+import { mockEndpoints, REQUEST_MEDIUM } from './helpers/mock-endpoints';
 import { withAnalyticsLogger, withIntlProvider } from './helpers';
 import AtlassianSwitcher from '../src';
 
-class GenericSwitcherExample extends Component {
+class GenericSwitcherExample extends React.Component {
   state = {
     isDrawerOpen: false,
   };
@@ -34,13 +34,7 @@ class GenericSwitcherExample extends Component {
           },
         };
       },
-      {
-        containers: 1000,
-        xflow: 500,
-        licenseInformation: 250,
-        permitted: 2000,
-        appswitcher: 1500,
-      },
+      REQUEST_MEDIUM,
     );
     this.setState({
       isDrawerOpen: true,