+++ packages/navigation/atlassian-switcher/src/providers/instance-data-providers.ts
@@ -1,68 +1,80 @@
 import { fetchJson, postJson } from '../utils/fetch';
-import asDataProvider, { DataProviderProps } from './as-data-provider';
-import { WithCloudId, RecentContainer } from '../types';
-import { LicenseInformationDataStructure } from './types';
+import asDataProvider from './as-data-provider';
+import {
+  LicenseInformationResponse,
+  Permissions,
+  RecentContainersResponse,
+  UserPermissionResponse,
+  WithCloudId,
+  XFlowSettingsResponse,
+} from '../types';
+import { withCached } from '../utils/with-cached';
 
-export interface CloudIdDataProvider<T>
-  extends DataProviderProps<T>,
-    WithCloudId {}
+// Recent activity api
+const fetchRecentContainers = ({ cloudId }: WithCloudId) =>
+  fetchJson<RecentContainersResponse>(
+    `/gateway/api/activity/api/client/recent/containers?cloudId=${cloudId}`,
+  );
 
-export interface RecentContainersDataStructure {
-  data: Array<RecentContainer>;
-}
+export const RecentContainersProvider = asDataProvider(fetchRecentContainers);
 
-export const RecentContainersProvider = asDataProvider(
-  ({ cloudId }: WithCloudId) =>
-    fetchJson<RecentContainersDataStructure>(
-      `/gateway/api/activity/api/client/recent/containers?cloudId=${cloudId}`,
-    ).then((recentContainersData: RecentContainersDataStructure) => ({
-      data: recentContainersData.data.slice(0, 6),
-    })),
+// License information api
+const fetchLicenseInformation = withCached(({ cloudId }: WithCloudId) =>
+  fetchJson<LicenseInformationResponse>(
+    `/gateway/api/xflow/${cloudId}/license-information`,
+  ),
 );
 
 export const LicenseInformationProvider = asDataProvider(
-  ({ cloudId }: WithCloudId) =>
-    fetchJson<LicenseInformationDataStructure>(
-      `/gateway/api/xflow/${cloudId}/license-information`,
-    ),
+  fetchLicenseInformation,
+  fetchLicenseInformation.cached,
 );
 
-export interface UserPermissionDataStructure {
-  permitted: boolean;
-}
-
-export enum Permissions {
-  MANAGE = 'manage',
-  CAN_INVITE_USERS = 'invite-users',
-  ADD_PRODUCTS = 'add-products',
-}
+// Permissions api
+type FetchPermissionParamsType = WithCloudId & {
+  permissionId: Permissions;
+};
+const fetchPermission = withCached(
+  ({ cloudId, permissionId }: FetchPermissionParamsType) =>
+    postJson<UserPermissionResponse>(`/gateway/api/permissions/permitted`, {
+      permissionId,
+      resourceId: `ari:cloud:platform::site/${cloudId}`,
+    }).then(permission => permission.permitted),
+);
 
 export const UserPermissionProvider = asDataProvider(
-  ({
-    cloudId,
-    permissionId,
-  }: WithCloudId & {
-    permissionId: Permissions;
-  }) =>
-    postJson<UserPermissionDataStructure>(
-      `/gateway/api/permissions/permitted`,
-      {
-        permissionId,
-        resourceId: `ari:cloud:platform::site/${cloudId}`,
-      },
-    ).then((permission): boolean => permission.permitted),
+  fetchPermission,
+  fetchPermission.cached,
 );
 
-export interface XFlowSettings {
-  [s: string]: any;
-}
+// Xflow settings api
+const fetchXflowSettings = withCached(({ cloudId }: WithCloudId) =>
+  fetchJson<XFlowSettingsResponse>(
+    `/gateway/api/site/${cloudId}/setting/xflow`,
+  ).then(xFlowSettings =>
+    xFlowSettings['product-suggestions-enabled'] !== undefined
+      ? xFlowSettings['product-suggestions-enabled']
+      : true,
+  ),
+);
 
 export const XFlowSettingsProvider = asDataProvider(
-  ({ cloudId }: WithCloudId) =>
-    fetchJson<XFlowSettings>(`/gateway/api/site/${cloudId}/setting/xflow`).then(
-      (xFlowSettings): boolean =>
-        xFlowSettings.hasOwnProperty('product-suggestions-enabled')
-          ? xFlowSettings['product-suggestions-enabled']
-          : true,
-    ),
+  fetchXflowSettings,
+  fetchXflowSettings.cached,
 );
+
+export const prefetchAll = ({ cloudId }: WithCloudId) => {
+  fetchLicenseInformation({ cloudId });
+  fetchXflowSettings({ cloudId });
+  fetchPermission({
+    cloudId,
+    permissionId: Permissions.ADD_PRODUCTS,
+  });
+  fetchPermission({ cloudId, permissionId: Permissions.MANAGE });
+};
+
+export const resetAll = () => {
+  fetchLicenseInformation.reset();
+  fetchXflowSettings.reset();
+  fetchPermission.reset();
+};