+++ packages/elements/reactions/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/reactions
 
+## 15.6.3
+- Updated dependencies [7261577953](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7261577953):
+  - @atlaskit/emoji@57.0.1
+  - @atlaskit/elements-test-helpers@0.3.0
+
 ## 15.6.2
 - Updated dependencies [7ab3e93996](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7ab3e93996):
   - @atlaskit/editor-test-helpers@7.0.6